requirements.txt
pandas
numpy
scikit-learn
matplotlib
scipy
dataclasses
pymongo
pyaml
certifi
pymongo[srv]==3.11
python-dotenv
fastapi
uvicorn 
#-e.
